requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
openai==1.3.0
langchain==0.1.0
faiss-cpu==1.13.2
python-dotenv==1.0.0
PyPDF2==4.0.1
python-docx==0.8.11